Move voice decoder
