+++ include/aud_int_internal.h
@@ -6,4 +6,4 @@ typedef struct
 
 extern aud_export_t gAudExports;
 
-#define META_AUDIO_VERSION "Meta Audio 0d2d2808"
+#define META_AUDIO_VERSION "Meta Audio Version: 92b7efb1ffebb94e3d20ed89f6dd72b50264b16f"
